Refactor subject
