+++ src/app/store/pages/pages.effects.ts
@@ -2,9 +2,9 @@ import { Injectable } from '@angular/core'
 import { Actions, Effect, ofType } from '@ngrx/effects'
 import { mergeMap } from 'rxjs/operators'
 
-import * as sensorsDataActions from '../../subject-page/store/sensors-data/sensors-data.actions'
-import * as sensorsActions from '../../subject-page/store/sensors/sensors.actions'
-import * as sourcesActions from '../../subject-page/store/sources/sources.actions'
+import * as sensorsDataActions from '../../subject/store/sensors-data/sensors-data.actions'
+import * as sensorsActions from '../../subject/store/sensors/sensors.actions'
+import * as sourcesActions from '../../subject/store/sources/sources.actions'
 import * as actions from './pages.actions'
 
 @Injectable()
