+++ src/app/subject/components/source-graphs/source-graph/source-graph.component.scss
