+++ src/app/subject/components/source-graphs/source-graphs.component.scss
