+++ src/app/subject/store/sensors-data/sensors-data.actions.ts
