+++ src/app/subject/store/sensors-data/sensors-data.reducer.ts
