+++ src/app/subject/store/sensors/sensors.actions.ts
