+++ src/app/subject/store/sensors/sensors.reducer.ts
