+++ src/app/subject/store/sources/sources.effects.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core'
 import { Actions, Effect, ofType } from '@ngrx/effects'
-import { Action } from '@ngrx/store'
-import { Observable, of } from 'rxjs'
+import { of } from 'rxjs'
 import { catchError, map, switchMap } from 'rxjs/operators'
 
 import { SourcesService } from '../../services/sources.service'
